// Dataset/Sample Codes/PowerAppsGridCustomizerControl/PAGridCustomizer/types.ts
export interface CellRendererProps {
  /** Raw value of the cell */
  readonly value: unknown;
  /** True when grid is in RTL mode */
  readonly isRTLMode?: boolean;
  /** Callback indicating the grid cell has been clicked */
  readonly onCellClicked?: (event?: React.MouseEvent<HTMLElement, MouseEvent> | MouseEvent) => void;
  /** Returns the validation error for the cell */
  readonly validationError?: Error | null;
  /** True when cell should be right aligned */
  readonly isRightAligned?: boolean;
  /** Callback to pragmatically start editing the cell */
  readonly startEditing?: (editorInitValue?: unknown) => void;
  /** True when cell is on the last row of the grid */
  readonly isLastRow?: boolean;
  /** Grid row height in pixels */
  readonly rowHeight?: number;
  /** Cell column data type */
  readonly columnDataType?: ColumnDataType;
  /** Formatted value of the cell */
  readonly formattedValue?: string;
  /** HTML element containing the cell */
  readonly cellContainerElement?: HTMLElement;
  /** Cell error label Id */
  readonly cellErrorLabelId?: string;
  /** True when the cell column is editable */
  readonly columnEditable?: boolean;
}

export interface GetRendererParams {
  /** Column definitions for all visible columns in the grid */
  readonly colDefs: ColumnDefinition[];
  /** Cell column index */
  readonly columnIndex: number;
  /** Cell row data */
  readonly rowData?: RowData;
  /** True when tab navigation is allowed in the grid */
  readonly allowTabKeyNavigation?: boolean;
}
// ...
export const RECID = '__rec_id';
export interface RowData {
  /** Unique id for the row */
  readonly [RECID]: string;
}

export interface ColumnDefinition {
  /** Field name (should be unique) */
  readonly name: string;
  /** Column display name. Will be shown as column header.If none provided shimmer will be shown in place of header */
  readonly displayName?: string;
  /** The unique ID to give the column. This is optional. If missing, the ID will default to the name.
   * If defined, this ID will be used to identify the column in the API for sorting and filtering. */
  readonly colId?: string;
  /** Data type for the values in the column. */
  readonly dataType: string;
  /** Custom Renderer type to be used for this column when custom renderer override is provided. */
  readonly CustomRendererType?: string;
  /** Width of the column. Auto calculated if not set. */
  readonly width?: number;
  /** Min width, in pixels, of the cell */
  readonly minWidth?: number;
  /** Max width, in pixels, of the cell */
  readonly maxWidth?: number;
  /** True if column is initially hidden
   * @Notice To show/hide columns after the first render use PAGridAPI.setColumnVisible()
   */
  readonly hide?: boolean;
  /** Column is primary field */
  readonly isPrimary: boolean;
}

export type ColumnDataType =
  | 'Text'
  | 'Email'
  | 'Phone'
  | 'Ticker'
  | 'URL'
  | 'TextArea'
  | 'Lookup'
  | 'Customer'
  | 'Owner'
  | 'MultiSelectPicklist'
  | 'OptionSet'
  | 'TwoOptions'
  | 'Duration'
  | 'Language'
  | 'Multiple'
  | 'TimeZone'
  | 'Integer'
  | 'Currency'
  | 'Decimal'
  | 'FloatingPoint'
  | 'AutoNumber'
  | 'DateOnly'
  | 'DateAndTime'
  | 'Image'
  | 'File'
  | 'Persona';